Code refactoring and few minor fixes
Last parser issue: incorrect token breakdown
