+++ Bondar_Kistaiev_FI-03/DataBase/DataBase.cpp
@@ -2,7 +2,7 @@
 
 query_result DataBase::select_from(const std::string& tbl, const std::string& cond)
 {
-    return {"SELECT CALLED ON TABLE:  " + tbl + "  with condition:  " + cond, Result_Code::TextMessage};
+    return {"SELECT CALLED ON TABLE:  " + tbl + "\nWith condition:  " + cond, Result_Code::TextMessage};
 }
 
 query_result DataBase::create_join(const std::string& t1,  const std::string& t2, const std::string& on_col1, const std::string& on_col2)
@@ -14,16 +14,27 @@ query_result DataBase::insert(const std::string& tbl, const std::vector<std::str
 {
     std::string temp = "(";
     for (auto s : row)
-        temp += (s + ", ");
-        
-    temp += ")";
+        temp += (s + ", "); 
+    temp.pop_back();
+    temp[temp.size() - 1] = ')';
 
-    return {"INSERT CALLED ON TABLE:  " + tbl + "  with values:  " + temp, Result_Code::TextMessage};
+    return {"INSERT CALLED ON TABLE:  " + tbl + "\nWith values:  " + temp, Result_Code::TextMessage};
 }
 
 query_result DataBase::create(const std::string& tbl_name, const std::vector<std::pair<std::string, bool>>& columns)
 {
-    return {"CREATED TABLE:  " + tbl_name, Result_Code::TextMessage};
+    std::string temp = "(";
+    for (auto [col, ind] : columns)
+    {
+        temp += col;
+        if(ind)
+            temp += " (indexed)";
+        temp += ", ";
+    }
+    temp.pop_back();
+    temp[temp.size() - 1] = ')';
+
+    return {"CREATED TABLE:  " + tbl_name + "\nWith columns: " + temp, Result_Code::TextMessage};
 }
                 
 query_result DataBase::get_table_string(const std::string& name) const
