+++ Bondar_Kistaiev_FI-03/main.cpp
@@ -1,9 +1,3 @@
-#include <iostream>
-#include <sstream>
-#include <string>
-#include <limits>
-#include <regex>
-#include <algorithm>
 #include "SystemControl\\SystemControl.hpp"
 
 int main()
